refactor query/mutations
